Refactor version module into smaller components
