+++ crates/core/src/version/tests/metadata.rs
@@ -1,4 +1,6 @@
-use super::{Brand, *};
+use super::{constants::sanitize_build_revision, *};
+use crate::{branding::Brand, workspace};
+use rsync_protocol::ProtocolVersion;
 
 #[test]
 fn version_metadata_matches_expected_constants() {
