+++ crates/core/src/version/tests/report.rs
@@ -1,4 +1,10 @@
-use super::{Brand, *};
+use super::*;
+use crate::branding::Brand;
+use crate::version::report::{
+    default_checksum_algorithms, default_compress_algorithms, default_daemon_auth_algorithms,
+};
+use libc::{ino_t, off_t, time_t};
+use std::mem;
 
 #[test]
 fn version_info_report_renders_default_report() {
